feat(menu): mobile support for options
this makes it so that when on a mobile device (or when simulating
mobile) that the menu options render as tappable buttons instead of the
keyboard-based input.

This approach could also be adapted for use with a mouse too on desktop!
So that's possible too but would require some modification.